This version loads the app successfully
