+++ .env
@@ -1,1 +1,1 @@
-DATABASE_URL=postgres://u99bjla610fucg:p5ed433c6f9b3651e2ce13811f9dc58d8925670d9fe4eb9e577de940566dc1ea9@cbdhrtd93854d5.cluster-czrs8kj4isg7.us-east-1.rds.amazonaws.com:5432/d7n50q788rl1m2
+export DATABASE_URL=postgresql://u99bjla610fucg:p5ed433c6f9b3651e2ce13811f9dc58d8925670d9fe4eb9e577de940566dc1ea9@cbdhrtd93854d5.cluster-czrs8kj4isg7.us-east-1.rds.amazonaws.com:5432/d7n50q788rl1m2
